fix numeric output from operator<<

+++ include/nic.h
@@ -57,7 +57,7 @@ public:
     friend OStream & operator << (OStream & cout, const Address & a) {
 	    cout << hex;
 	    for(unsigned int i = 0; i < LENGTH; i++) {
-		cout  << a._address[i];
+		cout  << (unsigned int)(a._address[i]);
 		if(i < (LENGTH - 1))
             cout << ":";
 	    }
@@ -69,7 +69,7 @@ public:
 	friend Debug & operator << (Debug & db, const Address & a) {
 	    db << hex;
 	    for(unsigned int i = 0; i < LENGTH; i++) {
-		db << a._address[i];
+		db << (unsigned int)(a._address[i]);
 		if(i < (LENGTH - 1))
 		    db << ":";
 	    }
